refactor(infra): generalize skill-upload to volume-upload and relocate system skill hash
- Rename skill-upload.ts to volume-upload.ts with storageName param instead of skillName
- Move computeSystemSkillHash() from infra/storage/content-hash.ts to zero/skills/content-hash.ts
- Update all callers to compute storage name themselves via getCustomSkillStorageName()
- Update knip.json entry and upload-storage.ts docstring reference

Closes #9672
